hrms/requirements.txt
django 
djangorestframework 
djangorestframework-simplejwt
django-cors-headers
psycopg2
psycopg2-binary
Pillow
weasyprint
PyPDF2
pdfplumber
PyMuPDF
python-dateutil
gunicorn
pandas
djangorestframework
drf-yasg
python-dotenv==0.21.0


